doubles.. are sadly not floats
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_0branch/bzflag@13223 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/Player.cxx
@@ -230,7 +230,7 @@ float Player::getMuzzleHeight() const
 void Player::move(const float* _pos, float _azimuth)
 {
 	// update the speed of the state
-	float currentTime = TimeKeeper::getCurrent().getSeconds();
+	float currentTime = (float)TimeKeeper::getCurrent().getSeconds();
 	if ( state.lastUpdateTime >= 0 )
 	{
 		float delta = currentTime - state.lastUpdateTime;
